+++ Modules/SharePointDsc/DSCResources/MSFT_SPUserProfileServiceApp/MSFT_SPUserProfileServiceApp.psm1
@@ -17,10 +17,6 @@ function Get-TargetResource
         $ApplicationPool,
 
         [parameter(Mandatory = $false)] 
-        [System.Management.Automation.PSCredential] 
-        $FarmAccount,
-
-        [parameter(Mandatory = $false)] 
         [System.String] 
         $MySiteHostLocation,
 
@@ -68,6 +64,44 @@ function Get-TargetResource
        
     Write-Verbose -Message "Getting user profile service application $Name"
 
+    $farmAccount = Invoke-SPDSCCommand -Credential $InstallAccount `
+                                       -Arguments $PSBoundParameters `
+                                       -ScriptBlock {
+        return Get-SPDSCFarmAccountName 
+    }
+
+    if ($null -ne $farmAccount)
+    {
+        if ($PSBoundParameters.ContainsKey("InstallAccount") -eq $true) 
+        {
+            # InstallAccount used
+            if ($InstallAccount.UserName -ne $farmAccount)
+            {
+                throw ("Specified InstallAccount ($($InstallAccount.UserName)) isn't the Farm " + `
+                       "Account. Make sure the specified InstallAccount is the Farm Account " + `
+                       "and try again")
+            }
+        }
+        else {
+            # PSDSCRunAsCredential or System
+            if (-not $Env:USERNAME.Contains("$"))
+            {
+                # PSDSCRunAsCredential used
+                $localaccount = "$($Env:USERDOMAIN)\$($Env:USERNAME)"
+                if ($localaccount -ne $farmAccount)
+                {
+                    throw ("Specified PSDSCRunAsCredential ($localaccount) isn't the Farm " + `
+                           "Account. Make sure the specified PSDSCRunAsCredential is the Farm " + `
+                           "Account and try again")
+                }
+            }
+        }
+    }
+    else
+    {
+        throw ("Unable to retrieve the Farm Account. Check if the farm exists.")
+    }
+
     $result = Invoke-SPDSCCommand -Credential $InstallAccount `
                                   -Arguments $PSBoundParameters `
                                   -ScriptBlock {
@@ -113,16 +147,6 @@ function Get-TargetResource
             }
             $databases.Add("SynchronizationDatabase", $syncProp.GetValue($serviceApp))
 
-            $spFarm = Get-SPFarm
-
-            if ($params.FarmAccount.UserName -eq $spFarm.DefaultServiceAccount.Name) 
-            {
-                $farmAccount = $params.FarmAccount
-            } 
-            else 
-            {
-                $farmAccount = $spFarm.DefaultServiceAccount.Name
-            }
             $serviceAppProxies = Get-SPServiceApplicationProxy -ErrorAction SilentlyContinue
             if ($null -ne $serviceAppProxies)
             {
@@ -138,7 +162,6 @@ function Get-TargetResource
                 Name               = $serviceApp.DisplayName
                 ProxyName          = $proxyName
                 ApplicationPool    = $serviceApp.ApplicationPool.Name
-                FarmAccount        = $farmAccount
                 MySiteHostLocation = $params.MySiteHostLocation
                 ProfileDBName      = $databases.ProfileDatabase.Name
                 ProfileDBServer    = $databases.ProfileDatabase.Server.Name
@@ -172,10 +195,6 @@ function Set-TargetResource
         [parameter(Mandatory = $true)] 
         [System.String] 
         $ApplicationPool,
-
-        [parameter(Mandatory = $false)] 
-        [System.Management.Automation.PSCredential] 
-        $FarmAccount,
 
         [parameter(Mandatory = $false)] 
         [System.String] 
@@ -227,24 +246,56 @@ function Set-TargetResource
 
     if ($Ensure -eq "Present") 
     {    
-        if ($PSBoundParameters.ContainsKey("FarmAccount") -eq $false) 
+        $farmAccount = Invoke-SPDSCCommand -Credential $InstallAccount `
+                                    -Arguments $PSBoundParameters `
+                                    -ScriptBlock {
+            return Get-SPDSCFarmAccountName
+        }
+
+        if ($null -ne $farmAccount)
         {
-            throw ("Unable to provision the user profile service without the Farm Account. " + `
-                   "Please specify the FarmAccount parameter and try again")
-            return
+            if ($PSBoundParameters.ContainsKey("InstallAccount") -eq $true) 
+            {
+                # InstallAccount used
+                if ($InstallAccount.UserName -ne $farmAccount)
+                {
+                    throw ("Specified InstallAccount ($($InstallAccount.UserName)) isn't the Farm Account. Make sure " + `
+                           "the specified InstallAccount is the Farm Account and try again")
+                }
+            }
+            else {
+                # PSDSCRunAsCredential or System
+                if (-not $Env:USERNAME.Contains("$"))
+                {
+                    # PSDSCRunAsCredential used
+                    $localaccount = "$($Env:USERDOMAIN)\$($Env:USERNAME)"
+                    if ($localaccount -ne $farmAccount)
+                    {
+                        throw ("Specified PSDSCRunAsCredential ($localaccount) isn't the Farm Account. Make sure " + `
+                               "the specified PSDSCRunAsCredential is the Farm Account and try again")
+                    }
+                }
+            }
+        }
+        else
+        {
+            throw ("Unable to retrieve the Farm Account. Check if the farm exists.")
         }
         
         Write-Verbose -Message "Creating user profile service application $Name"
         
-        # Add the FarmAccount to the local Administrators group, if it's not already there
-        $isLocalAdmin = Test-SPDSCUserIsLocalAdmin -UserName $FarmAccount.UserName
+        # Add the InstallAccount to the local Administrators group, if it's not already there
+        $isLocalAdmin = Test-SPDSCUserIsLocalAdmin -UserName $farmAccount
 
         if (!$isLocalAdmin)
         {
-            Add-SPDSCUserToLocalAdmin -UserName $FarmAccount.UserName
+            Add-SPDSCUserToLocalAdmin -UserName $farmAccount
+
+            # Cycle the Timer Service so that it picks up the local Admin token
+            Restart-Service -Name "SPTimerV4"
         }
 
-        $result = Invoke-SPDSCCommand -Credential $FarmAccount `
+        $result = Invoke-SPDSCCommand -Credential $InstallAccount `
                                       -Arguments $PSBoundParameters `
                                       -ScriptBlock {
             $params = $args[0]
@@ -273,7 +324,6 @@ function Set-TargetResource
             { 
                 $params.Remove("Ensure") | Out-Null 
             }
-            $params.Remove("FarmAccount") | Out-Null
 
             $params = Rename-SPDSCParamValue -params $params `
                                              -oldName "SyncDBName" `
@@ -324,10 +374,10 @@ function Set-TargetResource
             }
         }
 
-        # Remove the FarmAccount from the local Administrators group, if it was added above
+        # Remove the InstallAccount from the local Administrators group, if it was added above
         if (!$isLocalAdmin)
         {
-            Remove-SPDSCUserToLocalAdmin -UserName $FarmAccount.UserName
+            Remove-SPDSCUserToLocalAdmin -UserName $farmAccount
         }
     }
     
@@ -376,10 +426,6 @@ function Test-TargetResource
         [parameter(Mandatory = $true)] 
         [System.String] 
         $ApplicationPool,
-
-        [parameter(Mandatory = $false)] 
-        [System.Management.Automation.PSCredential] 
-        $FarmAccount,
 
         [parameter(Mandatory = $false)] 
         [System.String]