+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change log for SharePointDsc
 
+## 2.0
+
+The following changes will break 1.x configurations that use these resources:
+
+* Updated SPShellAdmin so it also works for non-content databases
+* Updated SPWebApplication so it infers the UseSSL value from the URL
+* Updated SPUSerProfileServiceApp and SPUserProfileSyncService to remove
+  the requirement for CredSSP
+
 ## 1.8
 
 * Fixed issue in SPServiceAppProxyGroup causing some service names to return as null